macros.d/vhost-rails.conf: remove PassengerInstanceRegistryDir setting.
Our intention with a per-vhost "PassengerInstanceRegistryDir" setting
was to place the registry information in a safer (but still writable
by the web server) location. However, newer versions of passenger
complain that the setting is not valid within a vhost:

  WARNING: PassengerInstanceRegistryDir cannot occur within
  <VirtualHost> section

Apparently, passenger has just been ignoring the repeated definitions,
and we've only gotten away with it for so long because we don't run
more than one rails site per physical server. The default behavior is
to use a subdirectory of the system's temporary directory (i.e. /tmp),
which -- if done correctly -- should be safe and work just fine for
us. So this commit removes the per-vhost settings and adopts the
default.